target-openrisc: use jmp_pc as flag variable for branches
At the moment a branch l.bf and l.bnf requires an additional flag variable
"btaken" for the jump to decide after the delayed slot
whether the jump should be taken or not.

With this patch the jmp_pc variable is used as a flag.
If jmp_pc is zero the branch is not taken.
This requires that together with the delayed slot flag
an additional branch flag is saved to differ between slot types.

In average the speed increases by around 5%.

Signed-off-by: Sebastian Macke <sebastian@macke.de>